% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmedio.R
\name{pmedio}
\alias{pmedio}
\title{Regla del punto medio}
\usage{
pmedio(f, a, b)
}
\arguments{
\item{f}{funcion que queremos integrar}

\item{a}{extremo izquierdo del intervalo donde queremos integrar}

\item{b}{extremo derecho del intervalo donde queremos integrar}
}
\value{
Devuelve la aproximacion de la integral de la funcion f entre a y b mediante la regla del punto medio
}
\description{
Este comando realiza una aproximacion de la integral de una funcion mediante la regla del punto medio
}
\examples{
pmedio(function(x){return(x^2)},3, 6)

}
\seealso{
\link{rectan} \link{trap} \link{simp}
}
