% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectan.R
\name{rectan}
\alias{rectan}
\title{Regla del rectangulo}
\usage{
rectan(f, a, b)
}
\arguments{
\item{f}{funcion que queremos integrar}

\item{a}{extremo izquierdo del intervalo donde queremos integrar}

\item{b}{extremo derecho del intervalo donde queremos integrar}
}
\value{
Devuelve la aproximacion de la integral de la funcion f entre a y b mediante la regla del rectangulo
}
\description{
Este comando realiza una aproximacion de la integral de una funcion mediante la regla del rectangulo
}
\examples{
rectan(function(x){return(x^2)},3, 5)
}
\seealso{
\link{pmedio} \link{trap} \link{simp}
}
