% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simp.R
\name{simp}
\alias{simp}
\title{Regla de Simpson}
\usage{
simp(f, a, b)
}
\arguments{
\item{f}{funcion que queremos integrar}

\item{a}{extremo izquierdo del intervalo donde queremos integrar}

\item{b}{extremo derecho del intervalo donde queremos integrar}
}
\value{
Devuelve la aproximacion de la integral de la funcion f entre a y b mediante la regla de Simpson
}
\description{
Este comando realiza una aproximacion de la integral de una funcion mediante la regla de Simpson
}
\examples{
simp(function(x){return(x^2)}, 2 ,5)

}
\seealso{
\link{rectan} \link{trap} \link{pmedio}
}
