% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trap.R
\name{trap}
\alias{trap}
\title{Regla del trapecio}
\usage{
trap(f, a, b)
}
\arguments{
\item{f}{funcion que queremos integrar}

\item{a}{extremo izquierdo del intervalo donde queremos integrar}

\item{b}{extremo derecho del intervalo donde queremos integrar}
}
\value{
Devuelve la aproximacion de la integral de la funcion f entre a y b mediante la regla del trapecio
}
\description{
Este comando realiza una aproximacion de la integral de una funcion mediante la regla del trapecio
}
\examples{
trap(function(x){return(x)}, 2 ,5)
}
\seealso{
\link{rectan} \link{pmedio} \link{simp}
}
